ADM-359 change test url
