+++ frontend/__tests__/src/components/metrics/ConfigStep/Board.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, within, screen } from '@testing-library/react'
 import { Board } from '@src/components/Metrics/ConfigStep/Board'
-import { BOARD_FIELDS, BOARD_TYPES, CONFIG_TITLE, ERROR_MESSAGE_COLOR } from '../../../fixtures'
+import { BOARD_FIELDS, BOARD_TYPES, CONFIG_TITLE, ERROR_MESSAGE_COLOR, MOCK_URL } from '../../../fixtures'
 import { Provider } from 'react-redux'
 import { setupStore } from '../../../utils/setupStoreUtil'
 import { setupServer } from 'msw/node'
@@ -34,7 +34,7 @@ const setup = () => {
   )
 }
 const server = setupServer(
-  rest.get('https://jsonplaceholder.typicode.com/posts', (req, res, ctx) => {
+  rest.get(MOCK_URL, (req, res, ctx) => {
     return res(ctx.status(200))
   })
 )
