+++ frontend/__tests__/src/fixtures.ts
@@ -43,3 +43,5 @@ export enum CONFIG_TITLE {
 }
 
 export const BOARD_FIELDS = ['board', 'boardId', 'email', 'projectKey', 'site', 'token']
+
+export const MOCK_URL = '/api/v1/jira/board'
